+++ actionview/lib/action_view/helpers/tags/collection_helpers.rb
@@ -19,6 +19,8 @@ module ActionView
 
           def label(label_html_options={}, &block)
             html_options = @input_html_options.slice(:index, :namespace).merge(label_html_options)
+            html_options[:for] ||= @input_html_options[:id] if @input_html_options[:id]
+
             @template_object.label(@object_name, @sanitized_attribute_name, @text, html_options, &block)
           end
         end